Rename `mem` to `Memory`
