+++ kernel/sys/tasks.cpp
@@ -12,7 +12,7 @@
 #include <Arch/Memory.hpp>
 #include <sys/tasks.hpp>
 #include <sys/Panic.hpp>
-#include <mem/heap.hpp>
+#include <Memory/heap.hpp>
 #include <Library/stdio.hpp>
 #include <Devices/serial/rs232.hpp>
 #include <stdint.h>
